+++ src/scrapeunblocker_haystack/__init__.py
@@ -1,4 +1,4 @@
-from scrapeunblocker_haystack.fetcher import ScrapeUnblockerFetcher
+from scrapeunblocker_haystack.fetcher import ScrapeUnblockerFetcher, StepExecutionError
 from scrapeunblocker_haystack.search import ScrapeUnblockerWebSearch
 
-__all__ = ["ScrapeUnblockerFetcher", "ScrapeUnblockerWebSearch"]
+__all__ = ["ScrapeUnblockerFetcher", "ScrapeUnblockerWebSearch", "StepExecutionError"]